v7.9 of DC's Website
Skipped around because this will be the last major version in v7, this update is huge. The SDKCommunity archive is done and all on one page but I also heavily moved around files to save a ton of space. Both in the links so its easier to send on discord/other sites and in the Neocities maximum of 500 files. Doing so has majorly helped to give room for the website to grow eventually.

I might create a game/app soon, idk though. That and I may put music because why not lol. Thats all for this update, thanks for reading lol.

Copyrighted material belong to respective owners, please give credit for any of my work you use.